Move protocol constants to API
