--- spigot/src/main/java/com/turikhay/mc/mapmodcompanion/spigot/ProtocolVersion.java
@@ -1,8 +1,0 @@
-package com.turikhay.mc.mapmodcompanion.spigot;
-
-public interface ProtocolVersion {
-    int MINECRAFT_1_13_2 = 404;
-    int MINECRAFT_1_14_4 = 498;
-    int MINECRAFT_1_15_2 = 578;
-    int MINECRAFT_1_16_3 = 753;
-}
